Added more player functionality, added config checks
* A function for checking the configuration options has been added.
* More functionality to the player has been added, they can now fire at
other gladiators but take no damage

+++ vars.c
@@ -82,7 +82,7 @@ bool verbose(verbosity_t v)
 static bool is_bool_valid(struct config_db *db)
 {
 	bool b = *(bool*)(db->addr);
-	return !(b > 1 || (!(db->zero_allowed) && !b));
+	return !((int)b > 1 || (!(db->zero_allowed) && !b));
 }
 
 static bool is_double_valid(struct config_db *db)
